requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
django-celery-beat==1.1.0
django-celery-results==1.0.1
django-simple-menu==1.2.1
ephem==3.7.6.0
future==0.16.0
idna==2.6
kombu==4.1.0
oauthlib==2.0.6
python-twitter==3.3
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22
vine==1.1.4
